+++ test/test_express_cliff.jl
@@ -46,15 +46,18 @@
         isequal(express(σᶻ, CR, UseOp), sZ)
     end
 
-    @testset "Clifford tensor operation apply regression" begin
+    @testset "Projector on stab state" begin
+        @test isequal(projector(Stabilizer(S"X")), projector(StabilizerState(S"X")))
+    end
+
+    @testset "Clifford tensor operation apply" begin
         gate = express(σˣ⊗σᶻ, CR, UseOp)
-        state = QuantumClifford.tensor(MixedDestabilizer(S"Z"), MixedDestabilizer(S"X"))
+        state = MixedDestabilizer(S"ZI IX")
         indices = [1, 2]
-        QuantumClifford.apply!(state, indices, gate)
+        apply!(state, indices, gate)
 
         @test isempty(indices)
-        @test QuantumClifford.expect(P"ZI", state) ≈ -1
-        @test QuantumClifford.expect(P"IX", state) ≈ -1
-        @test isequal(projector(Stabilizer(S"X")), projector(StabilizerState(S"X")))
+        @test expect(P"ZI", state) ≈ -1
+        @test expect(P"IX", state) ≈ -1
     end
 end